refactor: moved segments logic to separate feature directory
