+++ packages/react-native-enriched-markdown/android/src/main/baseline-prof.txt
@@ -145,7 +145,6 @@ Lcom/swmansion/enriched/markdown/renderer/BlockquoteRenderer;
 Lcom/swmansion/enriched/markdown/renderer/CodeBlockRenderer;
 Lcom/swmansion/enriched/markdown/renderer/CodeBlockRenderer$CodeBlockBoundaryPaddingSpan;
 Lcom/swmansion/enriched/markdown/renderer/CodeRenderer;
-Lcom/swmansion/enriched/markdown/renderer/DocumentRenderer;
 Lcom/swmansion/enriched/markdown/renderer/EmphasisRenderer;
 Lcom/swmansion/enriched/markdown/renderer/HeadingRenderer;
 Lcom/swmansion/enriched/markdown/renderer/HighlightRenderer;
@@ -255,22 +254,22 @@ Lcom/swmansion/enriched/markdown/styles/UnderlineStyle;
 # Utilities - Common
 Lcom/swmansion/enriched/markdown/utils/common/BreakStrategyUtils;
 Lcom/swmansion/enriched/markdown/utils/common/FeatureFlags;
-Lcom/swmansion/enriched/markdown/utils/common/MarkdownSegment;
-Lcom/swmansion/enriched/markdown/utils/common/MarkdownSegment$Math;
-Lcom/swmansion/enriched/markdown/utils/common/MarkdownSegment$Table;
-Lcom/swmansion/enriched/markdown/utils/common/MarkdownSegment$Text;
-Lcom/swmansion/enriched/markdown/utils/common/MarkdownSegmentKt;
-Lcom/swmansion/enriched/markdown/utils/common/MarkdownSegmentRenderer;
+Lcom/swmansion/enriched/markdown/segments/MarkdownSegment;
+Lcom/swmansion/enriched/markdown/segments/MarkdownSegment$Math;
+Lcom/swmansion/enriched/markdown/segments/MarkdownSegment$Table;
+Lcom/swmansion/enriched/markdown/segments/MarkdownSegment$Text;
+Lcom/swmansion/enriched/markdown/segments/MarkdownSegmentKt;
+Lcom/swmansion/enriched/markdown/segments/MarkdownSegmentRenderer;
 Lcom/swmansion/enriched/markdown/utils/common/MarkdownViewManagerUtilsKt;
 Lcom/swmansion/enriched/markdown/utils/common/MotionPreferencesKt;
 Lcom/swmansion/enriched/markdown/utils/common/ReadableMapExtensionsKt;
 Lcom/swmansion/enriched/markdown/utils/common/ReconciliationResult;
-Lcom/swmansion/enriched/markdown/utils/common/RenderedSegment;
-Lcom/swmansion/enriched/markdown/utils/common/RenderedSegment$Math;
-Lcom/swmansion/enriched/markdown/utils/common/RenderedSegment$Table;
-Lcom/swmansion/enriched/markdown/utils/common/RenderedSegment$Text;
-Lcom/swmansion/enriched/markdown/utils/common/SegmentReconciler;
-Lcom/swmansion/enriched/markdown/utils/common/SegmentSignature;
+Lcom/swmansion/enriched/markdown/segments/RenderedSegment;
+Lcom/swmansion/enriched/markdown/segments/RenderedSegment$Math;
+Lcom/swmansion/enriched/markdown/segments/RenderedSegment$Table;
+Lcom/swmansion/enriched/markdown/segments/RenderedSegment$Text;
+Lcom/swmansion/enriched/markdown/segments/SegmentReconciler;
+Lcom/swmansion/enriched/markdown/segments/SegmentSignature;
 Lcom/swmansion/enriched/markdown/utils/common/StreamingMarkdownFilter;
 Lcom/swmansion/enriched/markdown/utils/common/TableStreamingMode;
 Lcom/swmansion/enriched/markdown/utils/common/layout/RTLUtilsKt;
@@ -309,19 +308,19 @@ Lcom/swmansion/enriched/markdown/utils/text/view/TextSelectionColorsKt;
 Lcom/swmansion/enriched/markdown/utils/text/view/TextViewSetupKt;
 
 # Views
-Lcom/swmansion/enriched/markdown/views/BlockSegmentView;
+Lcom/swmansion/enriched/markdown/segments/BlockSegmentView;
 Lcom/swmansion/enriched/markdown/views/ContextMenuPopup;
 Lcom/swmansion/enriched/markdown/views/ContextMenuPopup$Builder;
 Lcom/swmansion/enriched/markdown/views/ContextMenuPopup$Icon;
 Lcom/swmansion/enriched/markdown/views/ContextMenuPopup$MenuItem;
-Lcom/swmansion/enriched/markdown/views/MathContainerView;
-Lcom/swmansion/enriched/markdown/views/TableContainerView;
-Lcom/swmansion/enriched/markdown/views/TableContainerView$CellBackgroundView;
-Lcom/swmansion/enriched/markdown/views/TableContainerView$CellTextView;
-Lcom/swmansion/enriched/markdown/views/TableContainerView$GridContainerView;
-Lcom/swmansion/enriched/markdown/views/TableContainerView$HeaderTypefaceSpan;
-Lcom/swmansion/enriched/markdown/views/TableContainerView$TableCellData;
+Lcom/swmansion/enriched/markdown/segments/MathContainerView;
+Lcom/swmansion/enriched/markdown/segments/TableContainerView;
+Lcom/swmansion/enriched/markdown/segments/TableContainerView$CellBackgroundView;
+Lcom/swmansion/enriched/markdown/segments/TableContainerView$CellTextView;
+Lcom/swmansion/enriched/markdown/segments/TableContainerView$GridContainerView;
+Lcom/swmansion/enriched/markdown/segments/TableContainerView$HeaderTypefaceSpan;
+Lcom/swmansion/enriched/markdown/segments/TableContainerView$TableCellData;
 
 # Math (optional, from src/math source set)
 Lcom/swmansion/enriched/markdown/spans/MathMeasureHelper;
-Lcom/swmansion/enriched/markdown/views/MathContainerView$RaTeXCanvasView;
+Lcom/swmansion/enriched/markdown/segments/MathContainerView$RaTeXCanvasView;
