+++ packages/react-native-enriched-markdown/android/src/main/java/com/swmansion/enriched/markdown/EnrichedMarkdown.kt
@@ -13,26 +13,26 @@ import com.facebook.react.uimanager.StateWrapper
 import com.swmansion.enriched.markdown.accessibility.AccessibilityLabels
 import com.swmansion.enriched.markdown.parser.Md4cFlags
 import com.swmansion.enriched.markdown.parser.Parser
+import com.swmansion.enriched.markdown.segments.BlockquoteContainerView
+import com.swmansion.enriched.markdown.segments.CodeBlockContainerView
+import com.swmansion.enriched.markdown.segments.ContainerNodeView
+import com.swmansion.enriched.markdown.segments.MarkdownSegmentRenderer
+import com.swmansion.enriched.markdown.segments.RenderedSegment
+import com.swmansion.enriched.markdown.segments.SegmentViewConfig
+import com.swmansion.enriched.markdown.segments.SegmentViewCreators
+import com.swmansion.enriched.markdown.segments.SegmentViewFactory
+import com.swmansion.enriched.markdown.segments.TableContainerView
+import com.swmansion.enriched.markdown.segments.splitASTIntoSegments
 import com.swmansion.enriched.markdown.spoiler.SpoilerOverlay
 import com.swmansion.enriched.markdown.styles.StyleConfig
 import com.swmansion.enriched.markdown.utils.common.BreakStrategyUtils
 import com.swmansion.enriched.markdown.utils.common.CodeBlockStreamingMode
-import com.swmansion.enriched.markdown.utils.common.MarkdownSegmentRenderer
-import com.swmansion.enriched.markdown.utils.common.RenderedSegment
 import com.swmansion.enriched.markdown.utils.common.StreamingMarkdownFilter
 import com.swmansion.enriched.markdown.utils.common.TableStreamingMode
 import com.swmansion.enriched.markdown.utils.common.isReducedMotionEnabled
-import com.swmansion.enriched.markdown.utils.common.splitASTIntoSegments
 import com.swmansion.enriched.markdown.utils.text.TailFadeInAnimator
 import com.swmansion.enriched.markdown.utils.text.view.SelectionMenuConfig
 import com.swmansion.enriched.markdown.utils.text.view.applySelectionColors
-import com.swmansion.enriched.markdown.views.BlockquoteContainerView
-import com.swmansion.enriched.markdown.views.CodeBlockContainerView
-import com.swmansion.enriched.markdown.views.ContainerNodeView
-import com.swmansion.enriched.markdown.views.SegmentViewConfig
-import com.swmansion.enriched.markdown.views.SegmentViewCreators
-import com.swmansion.enriched.markdown.views.SegmentViewFactory
-import com.swmansion.enriched.markdown.views.TableContainerView
 import java.util.EnumSet
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
